requirements.txt
google-genai
pyautogui
python-dotenv
sounddevice
numpy